Small updates
I'm beginning to make the graphics functions and event handling.
Currently the game is quit with ESCAPE.
In the future, 'runes' that are hovered by the mouse and are near your own will highlight.
Currenly they are highlighted with just clicking 1-6 and removing the highlight with 0.
The highlight does not spread to adjacent tiles of the same color, but that will hopefully be fixed soon

Also added a .gitignore